Version 2.0 complete

+++ server/controllers/listController.js
@@ -1,9 +1,10 @@
 const List = require("../models/listDB");
 
 const handleAddList = async (req, res) => {
+  const user = req.user;
   try {
     const { listName, listItems } = req.body;
-    const existingList = await List.findOne({ listName });
+    const existingList = await List.findOne({ listName, createdBy: user.id });
     if (existingList) {
       return res.status(400).json({ message: "List already exists!" });
     }
@@ -11,6 +12,7 @@ const handleAddList = async (req, res) => {
     const newList = await List.create({
       listName,
       listItems,
+      createdBy: user.id,
     });
 
     return res.status(201).json({
@@ -28,8 +30,9 @@ const handleAddList = async (req, res) => {
 };
 
 const handleGetAllLists = async (req, res) => {
+  const user = req.user;
   try {
-    const lists = await List.find();
+    const lists = await List.find({ createdBy: user.id });
     return res.status(200).send(lists);
   } catch (err) {
     console.error("Get all lists error:", err);
@@ -39,8 +42,9 @@ const handleGetAllLists = async (req, res) => {
 
 const handleGetListByID = async (req, res) => {
   const { id } = req.params;
+  const user = req.user;
   try {
-    const list = await List.findById(id);
+    const list = await List.findOne({ _id: id, createdBy: user.id });
     if (!list) return res.status(404).send("List not found");
     return res.status(200).send(list);
   } catch (err) {
@@ -51,9 +55,10 @@ const handleGetListByID = async (req, res) => {
 
 const handleUpdateList = async (req, res) => {
   const { id, listItems } = req.body;
+  const user = req.user;
 
   try {
-    const updatedList = await List.findById(id);
+    const updatedList = await List.findOne({ _id: id, createdBy: user.id });
     if (!updatedList) return res.status(404).send("List not found");
     updatedList.listItems = listItems;
     await updatedList.save();
@@ -65,12 +70,13 @@ const handleUpdateList = async (req, res) => {
 };
 
 const handleRemoveLists = async (req, res) => {
+  const user = req.user;
   try {
     const { listIds } = req.body;
     if (!listIds || listIds.length === 0) {
       return res.status(400).json({ message: "No lists selected" });
     }
-    const deletedLists = await List.deleteMany({ _id: { $in: listIds } });
+    const deletedLists = await List.deleteMany({ _id: { $in: listIds }, createdBy: user.id });
     return res.status(200).send(deletedLists);
   } catch (err) {
     console.error("Remove lists error:", err);
